back/requirements.txt
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==21.12b0
bleach==4.1.0
Brotli==1.0.9
cffi==1.15.0
click==8.0.3
cycler==0.11.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
fastapi==0.72.0
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.28.5
h11==0.12.0
httptools==0.3.0
idna==3.3
ipykernel==6.7.0
ipython==8.0.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multivolumefile==0.2.3
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
numpy==1.22.0
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
plotly==5.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
py7zr==0.17.2
pybcj==0.5.0
pycparser==2.21
pycryptodomex==3.12.0
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.6
pyppmd==0.17.3
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
pyzstd==0.15.1
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
stack-data==0.1.3
starlette==0.17.1
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
texttable==1.6.4
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
Werkzeug==2.0.2
scikit-image==0.19.1
python-docx==0.8.11
pytesseract==0.3.8